+++ packages/awino-ui/src/types/app.d.ts
@@ -20,7 +20,8 @@ export type I18nPageNamespace =
   | 'earn-liquidity-staking'
   | 'borrow'
   | 'borrow-details'
-  | 'swap';
+  | 'swap'
+  | 'podl';
 
 export type SetState<A> = React.Dispatch<React.SetStateAction<A>>;